hotfix-周三

+++ common/src/main/java/com/treefinance/saas/monitor/common/domain/dto/OperatorAllStatAccessDTO.java
@@ -58,37 +58,37 @@ public class OperatorAllStatAccessDTO implements Serializable {
     /**
      * 前n天的相同时刻创建任务数平均值
      */
-    private Integer previousEntryAvgCount;
+    private BigDecimal previousEntryAvgCount;
 
     /**
      * 前n天的相同时刻确认手机号数平均值
      */
-    private Integer previousConfirmMobileAvgCount;
+    private BigDecimal previousConfirmMobileAvgCount;
 
     /**
      * 前n天的相同时刻开始登陆数平均值
      */
-    private Integer previousStartLoginAvgCount;
+    private BigDecimal previousStartLoginAvgCount;
 
     /**
      * 前n天的相同时刻登陆成功数平均值
      */
-    private Integer previousLoginSuccessAvgCount;
+    private BigDecimal previousLoginSuccessAvgCount;
 
     /**
      * 前n天的相同时刻抓取成功数平均值
      */
-    private Integer previousCrawlSuccessAvgCount;
+    private BigDecimal previousCrawlSuccessAvgCount;
 
     /**
      * 前n天的相同时刻洗数成功数平均值
      */
-    private Integer previousProcessSuccessAvgCount;
+    private BigDecimal previousProcessSuccessAvgCount;
 
     /**
      * 前n天的相同时刻回调成功数平均值
      */
-    private Integer previousCallbackSuccessAvgCount;
+    private BigDecimal previousCallbackSuccessAvgCount;
 
     /**
      * 前n天的相同时刻确认手机转化率平均值
@@ -318,59 +318,59 @@ public class OperatorAllStatAccessDTO implements Serializable {
         this.dataType = dataType;
     }
 
-    public Integer getPreviousEntryAvgCount() {
+    public BigDecimal getPreviousEntryAvgCount() {
         return previousEntryAvgCount;
     }
 
-    public void setPreviousEntryAvgCount(Integer previousEntryAvgCount) {
+    public void setPreviousEntryAvgCount(BigDecimal previousEntryAvgCount) {
         this.previousEntryAvgCount = previousEntryAvgCount;
     }
 
-    public Integer getPreviousConfirmMobileAvgCount() {
+    public BigDecimal getPreviousConfirmMobileAvgCount() {
         return previousConfirmMobileAvgCount;
     }
 
-    public void setPreviousConfirmMobileAvgCount(Integer previousConfirmMobileAvgCount) {
+    public void setPreviousConfirmMobileAvgCount(BigDecimal previousConfirmMobileAvgCount) {
         this.previousConfirmMobileAvgCount = previousConfirmMobileAvgCount;
     }
 
-    public Integer getPreviousStartLoginAvgCount() {
+    public BigDecimal getPreviousStartLoginAvgCount() {
         return previousStartLoginAvgCount;
     }
 
-    public void setPreviousStartLoginAvgCount(Integer previousStartLoginAvgCount) {
+    public void setPreviousStartLoginAvgCount(BigDecimal previousStartLoginAvgCount) {
         this.previousStartLoginAvgCount = previousStartLoginAvgCount;
     }
 
-    public Integer getPreviousLoginSuccessAvgCount() {
+    public BigDecimal getPreviousLoginSuccessAvgCount() {
         return previousLoginSuccessAvgCount;
     }
 
-    public void setPreviousLoginSuccessAvgCount(Integer previousLoginSuccessAvgCount) {
+    public void setPreviousLoginSuccessAvgCount(BigDecimal previousLoginSuccessAvgCount) {
         this.previousLoginSuccessAvgCount = previousLoginSuccessAvgCount;
     }
 
-    public Integer getPreviousCrawlSuccessAvgCount() {
+    public BigDecimal getPreviousCrawlSuccessAvgCount() {
         return previousCrawlSuccessAvgCount;
     }
 
-    public void setPreviousCrawlSuccessAvgCount(Integer previousCrawlSuccessAvgCount) {
+    public void setPreviousCrawlSuccessAvgCount(BigDecimal previousCrawlSuccessAvgCount) {
         this.previousCrawlSuccessAvgCount = previousCrawlSuccessAvgCount;
     }
 
-    public Integer getPreviousProcessSuccessAvgCount() {
+    public BigDecimal getPreviousProcessSuccessAvgCount() {
         return previousProcessSuccessAvgCount;
     }
 
-    public void setPreviousProcessSuccessAvgCount(Integer previousProcessSuccessAvgCount) {
+    public void setPreviousProcessSuccessAvgCount(BigDecimal previousProcessSuccessAvgCount) {
         this.previousProcessSuccessAvgCount = previousProcessSuccessAvgCount;
     }
 
-    public Integer getPreviousCallbackSuccessAvgCount() {
+    public BigDecimal getPreviousCallbackSuccessAvgCount() {
         return previousCallbackSuccessAvgCount;
     }
 
-    public void setPreviousCallbackSuccessAvgCount(Integer previousCallbackSuccessAvgCount) {
+    public void setPreviousCallbackSuccessAvgCount(BigDecimal previousCallbackSuccessAvgCount) {
         this.previousCallbackSuccessAvgCount = previousCallbackSuccessAvgCount;
     }
 }
